requirements.txt
duckdb
pandas
pandas-ta
yfinance
requests
google-genai
python-dotenv
beautifulsoup4
